+++ src/exabgp/bgp/message/update/nlri/__init__.py
@@ -22,3 +22,4 @@ from exabgp.bgp.message.update.nlri.flow import Flow
 from exabgp.bgp.message.update.nlri.evpn import EVPN
 from exabgp.bgp.message.update.nlri.rtc import RTC
 from exabgp.bgp.message.update.nlri.bgpls import BGPLS
+from exabgp.bgp.message.update.nlri.mup import MUP